added sds-page__section created new sample pages, moved GSA specific samples to new GSA templates section
Extended sds-page with sds-page__section. sds-page is a wrapper to the midsection of a page beween the header-type sections and the footer and if primarily confers top and bottom padding. sds-page__section is a wrapper to each main column within the sds-page wrapper it removes the top margin from the first H tag in the wrapped element to ensure that things line up at the top